Move raw-public-key trust-stores to package rpkstore.
