+++ scandium-core/pom.xml
@@ -133,7 +133,8 @@
 							org.eclipse.californium.scandium.config,
 							org.eclipse.californium.scandium.dtls,
 							org.eclipse.californium.scandium.dtls.cipher,
-							org.eclipse.californium.scandium.dtls.pskstore
+							org.eclipse.californium.scandium.dtls.pskstore,
+							org.eclipse.californium.scandium.dtls.rpkstore
 						</Export-Package>
 						<Private-Package>
 							org.eclipse.californium.scandium.util
